html » json tile lists - prep for roster communication
* deleting karma and other dev tools

I do not use these tools, and they add to the serve time during at home development.

* jsonify appList

The applist was hard-coded into the dashboard,

It now can be seeded with json (or later perhaps via the database where apps present themselves to an api I've yet to make?)
The 'required conditions to show' does not work yet for player apps. :c

* e2e removal

e2e is part of protractor (a dev tester) which I do not use - and thus can be removed

* Update README.md

* removing test tools

further removal of test tools like karma and protractor - still had commands, to packages that were not present

* modern browsers

when this project was first 'created' with ng generate it was on angular 8 - we've moved on. And we expect modern browsers. Anyone on anything less, can upgrade.

This site is only required to function for modern devices

* json apps seeding

The apps are now seeded from a json file - the function to hide apps when the requirement isn't met doesn't exist yet

* app description for mobile

the app description now should show for mobile if pressed? Or selected? I don't know exactly. It works on pc, but probably not on a  phone. Anyway, there's an extra thing.

* Seeded from json & json old

Json old compatibility is returned
We now can try to see if we can communicate with roster